requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
beautifulsoup4==4.9.3
bleach==3.2.1
boto3==1.16.43
botocore==1.19.43
Cartridge==0.13.0
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==1.11.29
django-contrib-comments==1.9.2
django-storages==1.11.1
filebrowser-safe==0.5.0
future==0.18.2
grappelli-safe==0.5.2
html5lib==1.1
idna==2.10
jmespath==0.10.0
Mezzanine==4.3.1
oauthlib==3.1.0
packaging==20.8
Pillow==8.0.1
pkg-resources==0.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.4
reportlab==3.5.56
requests==2.25.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.2
webencodings==0.5.1
xhtml2pdf==0.2.5
